* logo in template

+++ app/model/file.php
@@ -38,11 +38,13 @@ class App_Model_File extends App_Model_Table
 		$file = $this->getRecord($file_id);
 
 		if ($file) {
-			if (file_exists($file['path'])) {
-				unlink($file['path']);
-			}
+			if ($file['customer_id'] === customer_info('customer_id') || user_can('w', 'admin/file')) {
+				if (file_exists($file['path'])) {
+					unlink($file['path']);
+				}
 
-			return parent::remove($file_id);
+				return parent::remove($file_id);
+			}
 		}
 
 		return false;
